refactor ui, storage, tasklis
