+++ src/main/java/DonkException.java
@@ -1,0 +1,6 @@
+public class DonkException extends Exception{
+
+    public DonkException(String msg) {
+        super(msg);
+    }
+}
